requirements.txt
Flask==3.0.0
mysql-connector-python==8.0.32
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.3.3
openpyxl==3.1.5
